+++ packages/clerk-js/src/core/resources/SignIn.ts
@@ -16,7 +16,6 @@ import type {
   EmailCodeConfig,
   EmailLinkConfig,
   EnterpriseSSOConfig,
-  OAuthStrategy,
   PassKeyConfig,
   PasskeyFactor,
   PhoneCodeConfig,
@@ -26,10 +25,18 @@ import type {
   ResetPasswordParams,
   ResetPasswordPhoneCodeFactorConfig,
   SamlConfig,
-  SetActiveNavigate,
   SignInCreateParams,
   SignInFirstFactor,
+  SignInFutureCreateParams,
+  SignInFutureEmailCodeSendParams,
+  SignInFutureEmailCodeVerifyParams,
+  SignInFutureFinalizeParams,
+  SignInFuturePasswordParams,
+  SignInFuturePhoneCodeSendParams,
+  SignInFuturePhoneCodeVerifyParams,
+  SignInFutureResetPasswordSubmitParams,
   SignInFutureResource,
+  SignInFutureSSOParams,
   SignInIdentifier,
   SignInJSON,
   SignInJSONSnapshot,
@@ -44,9 +51,11 @@ import type {
 } from '@clerk/types';
 
 import {
+  generateSignatureWithBase,
   generateSignatureWithCoinbaseWallet,
   generateSignatureWithMetamask,
   generateSignatureWithOKXWallet,
+  getBaseIdentifier,
   getCoinbaseWalletIdentifier,
   getMetamaskIdentifier,
   getOKXWalletIdentifier,
@@ -143,11 +152,8 @@ export class SignIn extends BaseResource implements SignInResource {
         } as PhoneCodeConfig;
         break;
       case 'web3_metamask_signature':
-        config = { web3WalletId: factor.web3WalletId } as Web3SignatureConfig;
-        break;
+      case 'web3_base_signature':
       case 'web3_coinbase_wallet_signature':
-        config = { web3WalletId: factor.web3WalletId } as Web3SignatureConfig;
-        break;
       case 'web3_okx_wallet_signature':
         config = { web3WalletId: factor.web3WalletId } as Web3SignatureConfig;
         break;
@@ -361,6 +367,15 @@ export class SignIn extends BaseResource implements SignInResource {
     });
   };
 
+  public authenticateWithBase = async (): Promise<SignInResource> => {
+    const identifier = await getBaseIdentifier();
+    return this.authenticateWithWeb3({
+      identifier,
+      generateSignature: generateSignatureWithBase,
+      strategy: 'web3_base_signature',
+    });
+  };
+
   public authenticateWithOKXWallet = async (): Promise<SignInResource> => {
     const identifier = await getOKXWalletIdentifier();
     return this.authenticateWithWeb3({
@@ -495,6 +510,11 @@ class SignInFuture implements SignInFutureResource {
     submitPassword: this.submitResetPassword.bind(this),
   };
 
+  phoneCode = {
+    sendCode: this.sendPhoneCode.bind(this),
+    verifyCode: this.verifyPhoneCode.bind(this),
+  };
+
   constructor(readonly resource: SignIn) {}
 
   get status() {
@@ -543,7 +563,8 @@ class SignInFuture implements SignInFutureResource {
     });
   }
 
-  async verifyResetPasswordEmailCode({ code }: { code: string }): Promise<{ error: unknown }> {
+  async verifyResetPasswordEmailCode(params: SignInFutureEmailCodeVerifyParams): Promise<{ error: unknown }> {
+    const { code } = params;
     return runAsyncResourceTask(this.resource, async () => {
       await this.resource.__internal_basePost({
         body: { code, strategy: 'reset_password_email_code' },
@@ -552,13 +573,8 @@ class SignInFuture implements SignInFutureResource {
     });
   }
 
-  async submitResetPassword({
-    password,
-    signOutOfOtherSessions = true,
-  }: {
-    password: string;
-    signOutOfOtherSessions?: boolean;
-  }): Promise<{ error: unknown }> {
+  async submitResetPassword(params: SignInFutureResetPasswordSubmitParams): Promise<{ error: unknown }> {
+    const { password, signOutOfOtherSessions = true } = params;
     return runAsyncResourceTask(this.resource, async () => {
       await this.resource.__internal_basePost({
         body: { password, signOutOfOtherSessions },
@@ -567,13 +583,7 @@ class SignInFuture implements SignInFutureResource {
     });
   }
 
-  async create(params: {
-    identifier?: string;
-    strategy?: OAuthStrategy | 'saml' | 'enterprise_sso';
-    redirectUrl?: string;
-    actionCompleteRedirectUrl?: string;
-    transfer?: boolean;
-  }): Promise<{ error: unknown }> {
+  async create(params: SignInFutureCreateParams): Promise<{ error: unknown }> {
     return runAsyncResourceTask(this.resource, async () => {
       await this.resource.__internal_basePost({
         path: this.resource.pathRoot,
@@ -582,7 +592,8 @@ class SignInFuture implements SignInFutureResource {
     });
   }
 
-  async password({ identifier, password }: { identifier?: string; password: string }): Promise<{ error: unknown }> {
+  async password(params: SignInFuturePasswordParams): Promise<{ error: unknown }> {
+    const { identifier, password } = params;
     return runAsyncResourceTask(this.resource, async () => {
       const previousIdentifier = this.resource.identifier;
       await this.resource.__internal_basePost({
@@ -592,7 +603,8 @@ class SignInFuture implements SignInFutureResource {
     });
   }
 
-  async sendEmailCode({ email }: { email: string }): Promise<{ error: unknown }> {
+  async sendEmailCode(params: SignInFutureEmailCodeSendParams): Promise<{ error: unknown }> {
+    const { email } = params;
     return runAsyncResourceTask(this.resource, async () => {
       if (!this.resource.id) {
         await this.create({ identifier: email });
@@ -612,7 +624,8 @@ class SignInFuture implements SignInFutureResource {
     });
   }
 
-  async verifyEmailCode({ code }: { code: string }): Promise<{ error: unknown }> {
+  async verifyEmailCode(params: SignInFutureEmailCodeVerifyParams): Promise<{ error: unknown }> {
+    const { code } = params;
     return runAsyncResourceTask(this.resource, async () => {
       await this.resource.__internal_basePost({
         body: { code, strategy: 'email_code' },
@@ -621,29 +634,49 @@ class SignInFuture implements SignInFutureResource {
     });
   }
 
-  async sso({
-    flow = 'auto',
-    strategy,
-    redirectUrl,
-    redirectUrlComplete,
-  }: {
-    flow?: 'auto' | 'modal';
-    strategy: OAuthStrategy | 'saml' | 'enterprise_sso';
-    redirectUrl: string;
-    redirectUrlComplete: string;
-  }): Promise<{ error: unknown }> {
+  async sendPhoneCode(params: SignInFuturePhoneCodeSendParams): Promise<{ error: unknown }> {
+    const { phoneNumber, channel = 'sms' } = params;
+    return runAsyncResourceTask(this.resource, async () => {
+      if (!this.resource.id) {
+        await this.create({ identifier: phoneNumber });
+      }
+
+      const phoneCodeFactor = this.resource.supportedFirstFactors?.find(f => f.strategy === 'phone_code');
+
+      if (!phoneCodeFactor) {
+        throw new Error('Phone code factor not found');
+      }
+
+      const { phoneNumberId } = phoneCodeFactor;
+      await this.resource.__internal_basePost({
+        body: { phoneNumberId, strategy: 'phone_code', channel },
+        action: 'prepare_first_factor',
+      });
+    });
+  }
+
+  async verifyPhoneCode(params: SignInFuturePhoneCodeVerifyParams): Promise<{ error: unknown }> {
+    const { code } = params;
+    return runAsyncResourceTask(this.resource, async () => {
+      await this.resource.__internal_basePost({
+        body: { code, strategy: 'phone_code' },
+        action: 'attempt_first_factor',
+      });
+    });
+  }
+
+  async sso(params: SignInFutureSSOParams): Promise<{ error: unknown }> {
+    const { flow = 'auto', strategy, redirectUrl, redirectCallbackUrl } = params;
     return runAsyncResourceTask(this.resource, async () => {
       if (flow !== 'auto') {
         throw new Error('modal flow is not supported yet');
       }
 
-      const redirectUrlWithAuthToken = SignIn.clerk.buildUrlWithAuth(redirectUrl);
-
       if (!this.resource.id) {
         await this.create({
           strategy,
-          redirectUrl: redirectUrlWithAuthToken,
-          actionCompleteRedirectUrl: redirectUrlComplete,
+          redirectUrl: SignIn.clerk.buildUrlWithAuth(redirectCallbackUrl),
+          actionCompleteRedirectUrl: redirectUrl,
         });
       }
 
@@ -655,7 +688,8 @@ class SignInFuture implements SignInFutureResource {
     });
   }
 
-  async finalize({ navigate }: { navigate?: SetActiveNavigate } = {}): Promise<{ error: unknown }> {
+  async finalize(params?: SignInFutureFinalizeParams): Promise<{ error: unknown }> {
+    const { navigate } = params || {};
     return runAsyncResourceTask(this.resource, async () => {
       if (!this.resource.createdSessionId) {
         throw new Error('Cannot finalize sign-in without a created session.');